refine(v26.2.1): close 3 final spec-compliance gaps
Audit pass uncovered three remaining gaps vs `specs/v26/V26_2_1_PLAN.md`:

1. **Migration script not deleted** (plan §3 PR #1: "deleted after
   PR #1 lands"). PR #1 has landed via cascade-merge; deleting now.
   `check_result_helpers.py` previously imported the parser from the
   migration script — those parser functions are now inlined into
   the gate so it's self-contained.

2. **No UTF-8 BOM E2E fixture** (plan §6 R3: "test includes UTF-8
   BOM fixture"). Added `corpora/fixtures/v26_2_1/struct_001_with_bom.tex`
   (3-byte BOM + missing-docclass body), a 6th E2E case in
   `test_rule_fix_integration.ml` documenting current "respect byte-0
   semantics" behaviour, and the fixture in
   `check_fix_integration_wired.py`'s expected set. BOM-aware
   insertion (place `\documentclass` AFTER BOM) is documented as
   v26.3 scope.

3. **`docs/MIGRATION_v26.2_to_v26.2.1.md`** updated to reflect the
   migration-script deletion (was: "available for reference"; now:
   "recoverable from PRs #265 / #270 history").

`corpora/fixtures/v26_2_1/README.md` table extended with the BOM
fixture and its expected behaviour.

Test results post-changes:
- `[fix-integration] PASS 6 cases` (was 5).
- `dune runtest latex-parse/src` fully green.
- `pre_release_check.py` 16/16 gates PASS.

Refs: `specs/v26/V26_2_1_PLAN.md` §3 PR #1, §6 R3.
