+++ corpora/fixtures/v26_2_1/struct_001_with_bom.tex
@@ -1,0 +1,1 @@
+﻿No docclass after BOM.
