+++ latex-perfectionist.opam
@@ -1,6 +1,6 @@
 # This file is generated by dune, edit dune-project instead
 opam-version: "2.0"
-version: "26.2.0"
+version: "26.2.1"
 synopsis: "LaTeX Perfectionist v25 bootstrap"
 maintainer: ["Clan Maintainers <maintainers@clan.dev>"]
 authors: ["Clan Labs"]
